Write dpmt_add_member() and dpmt_remove_member()
With keeping the list in descending order, the remove operation can be
easily operated at head.

The type of @member in memlis_t is fixed.

In our design, the usage of partcipants and departments are not
intuitive. Their name (int number) are recorded by their index in
struct matching and cannot be traced in their own struct.