Add hash for wchar string
Add search, insert operations. It won't need delete operation.

Since our data comes with the csv input, the initialization of whash is
written in incsv(), without an independent whash_init().
The design of incsv() is not confirmed. I intended to call incsv()
twice to parse the two input csv, but I needed to know the size of dpmt
and pacpt to allocate appropriate space for the whash structs.
Therefore, I merge the two parsing into incsv(), but then it is strange
and the problem of return appears.

whash_func() is originally designed for char string, maybe not a good
one for wchar.