Redesign incsv and fix whash operations
incsv:
Add an independent whash_init(), which uses the result from
count_line() and allocate for struct whash.
Then we can have an in-file parsing function for each csv.
The data in the hash table is given in these parsing functions.

whash_ops functions:
The return type of both search and deletion are uint32_t.

(struct wentry) .name and .number are also initialized in initializer.
We need them to identify a wentry is used or not. The checking of using
h->table + value is wrong and caused segmentation fault.

The assignment of @name is changed to wcsdup() from wcscpy() since it
is originally a NULL pointer.