Add output function and modify whash_search()
The return-type of whash_search() is modified to easily obtain both
name and number.

To output our algo result, we need a convenient method to access names
of entries. Thus we add @num2idx in struct whash to search in the hash
table by id as well, and add nameOf() to access the name string.

With these helps, result() is completed.